fix: ajustado alguns parametros para que seja possivel registrar a transerência do cartão de crédito

+++ api/src/main/java/com/devstoblu/banking_system/dtos/CreditCardRequestResponseDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record CreditCardRequestResponseDTO (
         String statusSolicitacao,
-        BigDecimal limiteAprovado,
+        Double limiteAprovado,
         String brand,
         String maskedPan,
         String cardToken,
